==> approach_10/14_train_test_split.ipynb <==
import pandas as pd

full_file_path = '../data/processed/export_combined_v1.csv'
df = pd.read_csv(full_file_path)

df.head()

from sklearn.model_selection import train_test_split

# Define the target variable
y = df["positionOrder"]
X = df.drop(columns=["positionOrder"])

# Perform stratified train-test split
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, shuffle=True, random_state=42
)

print(len(X_train[X_train["year"] >= 2018]), len(X_test[X_test["year"] >= 2018]))

# Display the shapes of the resulting datasets
print("X_train shape:", X_train.shape)
print("X_test shape:", X_test.shape)
print("y_train shape:", y_train.shape)
print("y_test shape:", y_test.shape)

import os

# Ensure the directory exists
os.makedirs('../data/train', exist_ok=True)
os.makedirs('../data/test', exist_ok=True)

X_validation.to_csv('../data/validation/validation.csv', index=False)
y_validation.to_csv('../data/validation/validation_labels.csv', index=False)

# Write train data to CSV
X_train.to_csv('../data/train/train.csv', index=False)
y_train.to_csv('../data/train/train_labels.csv', index=False)

# Write test data to CSV
X_test.to_csv('../data/test/test.csv', index=False)
y_test.to_csv('../data/test/test_labels.csv', index=False)
