==> approach_10/20_loading_data.ipynb <==
## Load Data

import pandas as pd

X_train = pd.read_csv('../data/train/train.csv')
y_train = pd.read_csv('../data/train/train_labels.csv')

X_test = pd.read_csv('../data/test/test.csv')
y_test = pd.read_csv('../data/test/test_labels.csv')

# Filter X_train and X_test for entries where the year is greater than 2018
X_train_2018 = X_train[X_train["year"] > 2018]
X_test_2018 = X_test[X_test["year"] > 2018]

# Filter y_train and y_test based on the same condition applied to X_train and X_test
y_train_2018 = y_train[X_train["year"] > 2018]
y_test_2018 = y_test[X_test["year"] > 2018]

# Print the shapes of the filtered datasets
print("X_train_2018 shape:", X_train_2018.shape)
print("X_test_2018 shape:", X_test_2018.shape)
print("y_train_2018 shape:", y_train_2018.shape)
print("y_test_2018 shape:", y_test_2018.shape)
